StackDB/pyproject.toml
[tool.poetry]
name = "stackdb"
version = "0.1.0"
description = ""
authors = ["stevensusas <hanqisu@sas.upenn.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^9.0.0"
pandas = "^2.2.3"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
